requirements.txt
requests
jinja2
moviepy
pillow
python-dotenv